pull out helm-project-operator controller setup code into pkg/setup
